requirements.txt
pyppeteer
mcp
python-dotenv
cloudinary
pydantic